refactor: move getter/setter/iterator to store.go and change some names more intuitive
