+++ x/fundraising/keeper/auction_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 func (suite *KeeperTestSuite) TestAuctionId() {
-	auctionId := suite.keeper.GetAuctionId(suite.ctx)
+	auctionId := suite.keeper.GetLastAuctionId(suite.ctx)
 	suite.Require().Equal(uint64(0), auctionId)
 
 	cacheCtx, _ := suite.ctx.CacheContext()
