+++ x/fundraising/keeper/bid.go
@@ -5,138 +5,15 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	gogotypes "github.com/gogo/protobuf/types"
 
 	"github.com/tendermint/fundraising/x/fundraising/types"
 )
 
-// GetSequence returns the last sequence number of the bid.
-func (k Keeper) GetSequence(ctx sdk.Context, auctionId uint64) uint64 {
-	var seq uint64
-	store := ctx.KVStore(k.storeKey)
-	bz := store.Get(types.GetSequenceKey(auctionId))
-	if bz == nil {
-		seq = 0 // initialize the sequence
-	} else {
-		val := gogotypes.UInt64Value{}
-		err := k.cdc.Unmarshal(bz, &val)
-		if err != nil {
-			panic(err)
-		}
-		seq = val.GetValue()
-	}
-	return seq
-}
-
 // GetNextSequence increments sequence number by one and set it.
 func (k Keeper) GetNextSequenceWithUpdate(ctx sdk.Context, auctionId uint64) uint64 {
-	seq := k.GetSequence(ctx, auctionId) + 1
+	seq := k.GetLastSequence(ctx, auctionId) + 1
 	k.SetSequence(ctx, auctionId, seq)
 	return seq
-}
-
-// SetSequence sets the sequence number for the auction.
-func (k Keeper) SetSequence(ctx sdk.Context, auctionId uint64, seq uint64) {
-	store := ctx.KVStore(k.storeKey)
-	bz := k.cdc.MustMarshal(&gogotypes.UInt64Value{Value: seq})
-	store.Set(types.GetSequenceKey(auctionId), bz)
-}
-
-// GetBid returns a bid for the given auction id and sequence number.
-// A bidder can have as many bids as they want, so sequence is required to get the bid.
-func (k Keeper) GetBid(ctx sdk.Context, auctionId uint64, sequence uint64) (bid types.Bid, found bool) {
-	store := ctx.KVStore(k.storeKey)
-	bz := store.Get(types.GetBidKey(auctionId, sequence))
-	if bz == nil {
-		return bid, false
-	}
-	k.cdc.MustUnmarshal(bz, &bid)
-	return bid, true
-}
-
-// SetBid sets a bid with the given arguments.
-func (k Keeper) SetBid(ctx sdk.Context, auctionId uint64, sequence uint64, bidderAcc sdk.AccAddress, bid types.Bid) {
-	store := ctx.KVStore(k.storeKey)
-	bz := k.cdc.MustMarshal(&bid)
-	store.Set(types.GetBidKey(auctionId, sequence), bz)
-	store.Set(types.GetBidIndexKey(bidderAcc, auctionId, sequence), []byte{})
-}
-
-// GetBids returns all bids registered in the store.
-func (k Keeper) GetBids(ctx sdk.Context) []types.Bid {
-	bids := []types.Bid{}
-	k.IterateBids(ctx, func(bid types.Bid) (stop bool) {
-		bids = append(bids, bid)
-		return false
-	})
-	return bids
-}
-
-// GetBidsByAuctionId returns all bids associated with the auction id that are registered in the store.
-func (k Keeper) GetBidsByAuctionId(ctx sdk.Context, auctionId uint64) []types.Bid {
-	bids := []types.Bid{}
-	k.IterateBidsByAuctionId(ctx, auctionId, func(bid types.Bid) (stop bool) {
-		bids = append(bids, bid)
-		return false
-	})
-	return bids
-}
-
-// GetBidsByBidder returns all bids associated with the bidder that are registered in the store.
-func (k Keeper) GetBidsByBidder(ctx sdk.Context, bidderAcc sdk.AccAddress) []types.Bid {
-	bids := []types.Bid{}
-	k.IterateBidsByBidder(ctx, bidderAcc, func(bid types.Bid) (stop bool) {
-		bids = append(bids, bid)
-		return false
-	})
-	return bids
-}
-
-// IterateBids iterates through all bids stored in the store and invokes callback function for each item.
-// Stops the iteration when the callback function returns true.
-func (k Keeper) IterateBids(ctx sdk.Context, cb func(bid types.Bid) (stop bool)) {
-	store := ctx.KVStore(k.storeKey)
-	iter := sdk.KVStorePrefixIterator(store, types.BidKeyPrefix)
-	defer iter.Close()
-	for ; iter.Valid(); iter.Next() {
-		var bid types.Bid
-		k.cdc.MustUnmarshal(iter.Value(), &bid)
-		if cb(bid) {
-			break
-		}
-	}
-}
-
-// IterateBidsByAuctionId iterates through all bids associated with the auction id stored in the store
-// and invokes callback function for each item.
-// Stops the iteration when the callback function returns true.
-func (k Keeper) IterateBidsByAuctionId(ctx sdk.Context, auctionId uint64, cb func(bid types.Bid) (stop bool)) {
-	store := ctx.KVStore(k.storeKey)
-	iter := sdk.KVStorePrefixIterator(store, types.GetBidAuctionIDKey(auctionId))
-	defer iter.Close()
-	for ; iter.Valid(); iter.Next() {
-		var bid types.Bid
-		k.cdc.MustUnmarshal(iter.Value(), &bid)
-		if cb(bid) {
-			break
-		}
-	}
-}
-
-// IterateBidsByBidder iterates through all bids associated with the bidder stored in the store
-// and invokes callback function for each item.
-// Stops the iteration when the callback function returns true.
-func (k Keeper) IterateBidsByBidder(ctx sdk.Context, bidderAcc sdk.AccAddress, cb func(bid types.Bid) (stop bool)) {
-	store := ctx.KVStore(k.storeKey)
-	iter := sdk.KVStorePrefixIterator(store, types.GetBidIndexByBidderPrefix(bidderAcc))
-	defer iter.Close()
-	for ; iter.Valid(); iter.Next() {
-		auctionId, sequence := types.ParseBidIndexKey(iter.Key())
-		bid, _ := k.GetBid(ctx, auctionId, sequence)
-		if cb(bid) {
-			break
-		}
-	}
 }
 
 // ReservePayingCoin reserves paying coin to the paying reserve account.
